Fix: Cane of Byrna Cleanup Killing SFX30
Changelog: Fixed Byrna Cleanup Routine killing SFX 30. This was caused by
`last_cane_of_byrna_item_id` never being able to be -1 because it was declared
as byte, instead of int.

+++ src/link.h
@@ -234,7 +234,7 @@ class LinkClass : public sprite
     bool dontdraw;
     bool diagonalMovement;
     bool bigHitbox;
-    byte last_cane_of_byrna_item_id; 
+    int last_cane_of_byrna_item_id; 
     
 	int last_lens_id;// The item ID of the last Lens of Truth type item used
